+++ ui/app/components/app/app-header/app-header.component.js
@@ -101,8 +101,6 @@ export default class AppHeader extends PureComponent {
             <img
               className="app-header__metafox-logo app-header__metafox-logo--icon"
               src="/images/logo/metamask-fox.svg"
-              height={42}
-              width={42}
             />
           </div>
           <div className="app-header__account-menu-container">